generalize html accessibility test
