+++ test/utils/index.ts
@@ -4,5 +4,4 @@ export { default as assertWithTimeout } from './assertWithTimeout'
 export { default as consoleUtil } from './consoleUtil'
 export { default as domEvent } from './domEvent'
 export { default as syntheticEvent } from './syntheticEvent'
-export { default as expectAxe } from './expectAxe'
 export * from './testingComponents'
